refactor: 컨벤션 준수를 위한 코드 리팩토링
ALL 코드 컨벤션 맞춤
test.domain.car CarsTest,PositionTest를 분리하여 저장
