+++ src/main/java/racinggame/domain/car/Car.java
@@ -2,12 +2,6 @@ package racinggame.domain.car;
 
 import racinggame.domain.data.GameStatus;
 
-/**
- * Car 클래스는 자동차 경주에 사용될 자동차의 정보값을 담는 클래스이다.
- * 현재 위치와 이름 값을 필드에 보유하고 있으며,
- * 추가로 static 변수인 winnerRecord을 가지고 있어
- * Car 클래스가 보유한 position값 중 제일 큰 position 값이 무엇인지 조회할 수 있다.
- */
 public class Car {
     private final String name;
     private Position position;
@@ -25,6 +19,7 @@ public class Car {
         this.name = name;
         this.position = new Position(position);
     }
+
     public void accelerate() {
         position.accelerate();
     }
@@ -33,11 +28,6 @@ public class Car {
         return this.position.match(position);
     }
 
-    /**
-     * passingLog는 GameStatus 인스턴스를 전달받은 후, 현재 이 Car 인스턴스의 상태 기록을 업데이트하는 메서드이다.
-     *
-     * @param gameStatus 게임 상태값을 저장하는 Output 인스턴스이다.
-     */
     public void passingLog(GameStatus gameStatus) {
         gameStatus.updateCarStatus(name, position.getPosition());
     }
