+++ src/test/java/racinggame/domain/data/GameStatusTest.java
@@ -3,8 +3,6 @@ package racinggame.domain.data;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import racinggame.domain.car.Car;
 
 import java.util.ArrayList;
@@ -15,14 +13,14 @@ public class GameStatusTest {
     Names names;
 
     @BeforeEach
-    void init(){
+    void init() {
         names = new Names("pobi,crong,honux");
         gameStatus = new GameStatus(names.splitNamesByComma());
     }
 
     @Test
     void 우승자_확인_테스트() {
-        List<Car> carList= new ArrayList<>();
+        List<Car> carList = new ArrayList<>();
         carList.add(new Car("pobi", 0));
         carList.add(new Car("crong", 0));
         carList.add(new Car("honux", 0));
