+++ src/test/java/racinggame/domain/game/RacingGameTest.java
@@ -13,12 +13,13 @@ public class RacingGameTest {
     Names names = new Names("pobi,crong,honux");
     Repeat repeat = new Repeat("5");
     RacingGame racingGame = new RacingGame(names, new TestMovableStrategy());
+
     /**
      * 임의로 모든 차가 전진만 100% 하도록 인터페이스를 변경하여 생성하였다.
      * 그렇기 떄문에, 실제 이 게임의 경우 세 플레이어 모두 전진하여, 셋 모두 우승자임이 분명해진다.
      */
     @Test
-    void 레이싱게임_이동_테스트(){
+    void 레이싱게임_이동_테스트() {
         Map<String, Integer> gameLog;
         racingGame.moveCars();
         gameLog = racingGame.moveCars();
@@ -26,6 +27,7 @@ public class RacingGameTest {
         Assertions.assertThat(gameLog.get("pobi")).isEqualTo(2);
         Assertions.assertThat(gameLog.get("pobi")).isEqualTo(2);
     }
+
     @Test
     void 레이싱게임_승자_테스트() {
         for (int repeatIterator = 0; repeat.isLoopDone(repeatIterator); repeatIterator++) {
